requirements.txt
# Clean Requirements.txt for PythonAnywhere Deployment
# Tender Management System

# Core Flask Framework
Flask==3.1.1
Werkzeug==3.1.3

# Database & ORM
SQLAlchemy==2.0.36
Flask-SQLAlchemy==3.1.1
PyMySQL==1.1.1

# Database Migrations
Flask-Migrate==4.1.0

# Authentication
Flask-Login==0.6.3

# Configuration
python-dotenv==1.1.0

# PDF Generation
reportlab==4.4.1

# Excel File Generation
XlsxWriter==3.2.3

# Timezone Support
tzlocal==5.3.1

# Production WSGI Server (for PythonAnywhere)
gunicorn==21.2.0